// src/backend/generated/models/model_PlayerBuildRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { model_TowerWordRequest } from "./model_TowerWordRequest";
export type model_PlayerBuildRequest = {
  done?: boolean;
  words?: Array<model_TowerWordRequest>;
};