// src/backend/generated/models/model_PlayerExtendedWordsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type model_PlayerExtendedWordsResponse = {
  mapSize?: Array<number>;
  nextTurnSec?: number;
  roundEndsAt?: string;
  shuffleLeft?: number;
  turn?: number;
  usedIndexes?: Array<number>;
  words?: Array<string>;
};